*** empty log message ***
git-svn-id: https://ncisvn.nci.nih.gov/svn/cagrid/trunk@6870 dd7b0c16-3c94-4492-92de-414b6a06f0b1

+++ cagrid-1-0/caGrid/projects/gts/src/gov/nih/nci/cagrid/gts/service/PermissionManager.java
@@ -20,6 +20,7 @@ import java.util.List;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 
+
 /**
  * @author <A HREF="MAILTO:langella@bmi.osu.edu">Stephen Langella </A>
  * @author <A HREF="MAILTO:oster@bmi.osu.edu">Scott Oster </A>
@@ -37,14 +38,15 @@ public class PermissionManager {
 
 	private Database db;
 
+
 	public PermissionManager(DBManager dbManager) {
 		log = LogFactory.getLog(this.getClass().getName());
 		this.dbManager = dbManager;
 		this.db = dbManager.getDatabase();
 	}
 
-	public synchronized void addPermission(Permission p)
-			throws GTSInternalFault, IllegalPermissionFault {
+
+	public synchronized void addPermission(Permission p) throws GTSInternalFault, IllegalPermissionFault {
 
 		// This method assumes that any Trusted Authorites associated with a
 		// permission is valid
@@ -55,49 +57,41 @@ public class PermissionManager {
 
 		if (p.getGridIdentity() == null) {
 			IllegalPermissionFault fault = new IllegalPermissionFault();
-			fault.setFaultString("The permission " + formatPermission(p)
-					+ " no grid identity specified.");
+			fault.setFaultString("The permission " + formatPermission(p) + " no grid identity specified.");
 			throw fault;
 		}
 
 		if (p.getRole() == null) {
 			IllegalPermissionFault fault = new IllegalPermissionFault();
-			fault.setFaultString("The permission " + formatPermission(p)
-					+ " no role specified.");
+			fault.setFaultString("The permission " + formatPermission(p) + " no role specified.");
 			throw fault;
 		}
 
-		if ((p.getTrustedAuthorityName()
-				.equals(Constants.ALL_TRUST_AUTHORITIES))
-				&& (!p.getRole().equals(Role.TrustServiceAdmin))) {
+		if ((p.getTrustedAuthorityName().equals(Constants.ALL_TRUST_AUTHORITIES))
+			&& (!p.getRole().equals(Role.TrustServiceAdmin))) {
 			IllegalPermissionFault fault = new IllegalPermissionFault();
-			fault.setFaultString("The permission " + formatPermission(p)
-					+ " must specify a specific Trust Authority.");
+			fault.setFaultString("The permission " + formatPermission(p) + " must specify a specific Trust Authority.");
 			throw fault;
 		}
-		if ((!p.getTrustedAuthorityName().equals(
-				Constants.ALL_TRUST_AUTHORITIES))
-				&& (p.getRole().equals(Role.TrustServiceAdmin))) {
+		if ((!p.getTrustedAuthorityName().equals(Constants.ALL_TRUST_AUTHORITIES))
+			&& (p.getRole().equals(Role.TrustServiceAdmin))) {
 			IllegalPermissionFault fault = new IllegalPermissionFault();
 			fault.setFaultString("The permission " + formatPermission(p)
-					+ " cannot specify a specific Trust Authority.");
+				+ " cannot specify a specific Trust Authority.");
 			throw fault;
 		}
 
 		if (this.doesPermissionExist(p)) {
 			IllegalPermissionFault fault = new IllegalPermissionFault();
-			fault.setFaultString("The permission " + formatPermission(p)
-					+ " cannot be added, it already exists.");
+			fault.setFaultString("The permission " + formatPermission(p) + " cannot be added, it already exists.");
 			throw fault;
 		}
 
 		StringBuffer insert = new StringBuffer();
 		Connection c = null;
 		try {
-			insert.append("INSERT INTO " + PermissionsTable.TABLE_NAME
-					+ " SET " + PermissionsTable.GRID_IDENTITY + "= ?,"
-					+ PermissionsTable.ROLE + "= ?,"
-					+ PermissionsTable.TRUSTED_AUTHORITY + "= ?");
+			insert.append("INSERT INTO " + PermissionsTable.TABLE_NAME + " SET " + PermissionsTable.GRID_IDENTITY
+				+ "= ?," + PermissionsTable.ROLE + "= ?," + PermissionsTable.TRUSTED_AUTHORITY + "= ?");
 			c = db.getConnection();
 			PreparedStatement s = c.prepareStatement(insert.toString());
 			s.setString(1, p.getGridIdentity());
@@ -106,35 +100,27 @@ public class PermissionManager {
 			s.execute();
 			s.close();
 		} catch (Exception e) {
-			this.log
-					.error(
-							"Unexpected database error incurred in adding the permission "
-									+ formatPermission(p)
-									+ ", the following statement generated the error: \n"
-									+ insert.toString() + "\n", e);
+			this.log.error("Unexpected database error incurred in adding the permission " + formatPermission(p)
+				+ ", the following statement generated the error: \n" + insert.toString() + "\n", e);
 			GTSInternalFault fault = new GTSInternalFault();
-			fault.setFaultString("Unexpected error adding the permission "
-					+ formatPermission(p) + "!!!");
+			fault.setFaultString("Unexpected error adding the permission " + formatPermission(p) + "!!!");
 			throw fault;
 		} finally {
 			db.releaseConnection(c);
 		}
 	}
 
-	public synchronized void revokePermission(Permission p)
-			throws GTSInternalFault, InvalidPermissionFault {
+
+	public synchronized void revokePermission(Permission p) throws GTSInternalFault, InvalidPermissionFault {
 		buildDatabase();
 		if (!doesPermissionExist(p)) {
 			InvalidPermissionFault fault = new InvalidPermissionFault();
-			fault.setFaultString("Could not revoke " + formatPermission(p)
-					+ ", the permission does not exist!!!");
+			fault.setFaultString("Could not revoke " + formatPermission(p) + ", the permission does not exist!!!");
 			throw fault;
 		}
 
-		String sql = "delete from " + PermissionsTable.TABLE_NAME + " where "
-				+ PermissionsTable.GRID_IDENTITY + "= ? AND "
-				+ PermissionsTable.ROLE + "= ? AND "
-				+ PermissionsTable.TRUSTED_AUTHORITY + "= ?";
+		String sql = "delete from " + PermissionsTable.TABLE_NAME + " where " + PermissionsTable.GRID_IDENTITY
+			+ "= ? AND " + PermissionsTable.ROLE + "= ? AND " + PermissionsTable.TRUSTED_AUTHORITY + "= ?";
 		Connection c = null;
 		try {
 			c = db.getConnection();
@@ -146,15 +132,10 @@ public class PermissionManager {
 			s.close();
 		} catch (Exception e) {
 			String perm = formatPermission(p);
-			this.log
-					.error(
-							"Unexpected database error incurred in removing the permission "
-									+ perm
-									+ " exists, the following statement generated the error: \n"
-									+ sql + "\n", e);
+			this.log.error("Unexpected database error incurred in removing the permission " + perm
+				+ " exists, the following statement generated the error: \n" + sql + "\n", e);
 			GTSInternalFault fault = new GTSInternalFault();
-			fault.setFaultString("Unexpected error in removing the permission "
-					+ perm + " exists.");
+			fault.setFaultString("Unexpected error in removing the permission " + perm + " exists.");
 			throw fault;
 		} finally {
 			db.releaseConnection(c);
@@ -162,12 +143,10 @@ public class PermissionManager {
 
 	}
 
-	public synchronized boolean doesPermissionExist(Permission p)
-			throws GTSInternalFault {
-		String sql = "select count(*) from " + PermissionsTable.TABLE_NAME
-				+ " where " + PermissionsTable.GRID_IDENTITY + "= ? AND "
-				+ PermissionsTable.ROLE + "= ?  AND "
-				+ PermissionsTable.TRUSTED_AUTHORITY + "= ?";
+
+	public synchronized boolean doesPermissionExist(Permission p) throws GTSInternalFault {
+		String sql = "select count(*) from " + PermissionsTable.TABLE_NAME + " where " + PermissionsTable.GRID_IDENTITY
+			+ "= ? AND " + PermissionsTable.ROLE + "= ?  AND " + PermissionsTable.TRUSTED_AUTHORITY + "= ?";
 		Connection c = null;
 		boolean exists = false;
 		try {
@@ -187,16 +166,10 @@ public class PermissionManager {
 			s.close();
 		} catch (Exception e) {
 			String perm = formatPermission(p);
-			this.log
-					.error(
-							"Unexpected database error incurred in determining if the permission "
-									+ perm
-									+ " exists, the following statement generated the error: \n"
-									+ sql + "\n", e);
+			this.log.error("Unexpected database error incurred in determining if the permission " + perm
+				+ " exists, the following statement generated the error: \n" + sql + "\n", e);
 			GTSInternalFault fault = new GTSInternalFault();
-			fault
-					.setFaultString("Unexpected error in determining if the permission "
-							+ perm + " exists.");
+			fault.setFaultString("Unexpected error in determining if the permission " + perm + " exists.");
 			throw fault;
 		} finally {
 			db.releaseConnection(c);
@@ -204,18 +177,16 @@ public class PermissionManager {
 		return exists;
 	}
 
-	public boolean isUserTrustServiceAdmin(String gridIdentity)
-			throws GTSInternalFault {
+
+	public boolean isUserTrustServiceAdmin(String gridIdentity) throws GTSInternalFault {
 		this.buildDatabase();
 		Connection c = null;
 		boolean isAdmin = false;
 		StringBuffer sql = new StringBuffer();
 		sql.append("select count(*) from " + PermissionsTable.TABLE_NAME);
 		sql.append(" WHERE " + PermissionsTable.GRID_IDENTITY + " = ? AND ");
-		sql.append(PermissionsTable.ROLE + "='" + Role.TrustServiceAdmin
-				+ "' AND ");
-		sql.append(PermissionsTable.TRUSTED_AUTHORITY + " = '"
-				+ Constants.ALL_TRUST_AUTHORITIES + "'");
+		sql.append(PermissionsTable.ROLE + "='" + Role.TrustServiceAdmin + "' AND ");
+		sql.append(PermissionsTable.TRUSTED_AUTHORITY + " = '" + Constants.ALL_TRUST_AUTHORITIES + "'");
 		try {
 			c = db.getConnection();
 			PreparedStatement s = c.prepareStatement(sql.toString());
@@ -231,17 +202,12 @@ public class PermissionManager {
 			s.close();
 
 		} catch (Exception e) {
-			this.log
-					.error(
-							"Unexpected database error incurred in determining whether or not the user "
-									+ gridIdentity
-									+ "  is a trust service administrator, the following statement generated the error: \n"
-									+ sql.toString() + "\n", e);
+			this.log.error("Unexpected database error incurred in determining whether or not the user " + gridIdentity
+				+ "  is a trust service administrator, the following statement generated the error: \n"
+				+ sql.toString() + "\n", e);
 			GTSInternalFault fault = new GTSInternalFault();
-			fault
-					.setFaultString("Unexpected error occurred in determining whether or not the user "
-							+ gridIdentity
-							+ "  is a trust service administrator.");
+			fault.setFaultString("Unexpected error occurred in determining whether or not the user " + gridIdentity
+				+ "  is a trust service administrator.");
 			throw fault;
 		} finally {
 			db.releaseConnection(c);
@@ -249,19 +215,16 @@ public class PermissionManager {
 		return isAdmin;
 	}
 
-	public boolean isUserTrustedAuthorityAdmin(String authority,
-			String gridIdentity) throws GTSInternalFault {
+
+	public boolean isUserTrustedAuthorityAdmin(String authority, String gridIdentity) throws GTSInternalFault {
 		this.buildDatabase();
 		Connection c = null;
 		boolean isAdmin = false;
 		StringBuffer sql = new StringBuffer();
 		sql.append("select count(*) from " + PermissionsTable.TABLE_NAME);
-		sql.append(" WHERE " + PermissionsTable.GRID_IDENTITY + " = ?"
-				+ " AND ");
-		sql.append(PermissionsTable.ROLE + "='" + Role.TrustAuthorityManager
-				+ "' AND ");
-		sql.append(PermissionsTable.TRUSTED_AUTHORITY + " = '" + authority
-				+ "'");
+		sql.append(" WHERE " + PermissionsTable.GRID_IDENTITY + " = ?" + " AND ");
+		sql.append(PermissionsTable.ROLE + "='" + Role.TrustAuthorityManager + "' AND ");
+		sql.append(PermissionsTable.TRUSTED_AUTHORITY + " = '" + authority + "'");
 		try {
 			c = db.getConnection();
 			PreparedStatement s = c.prepareStatement(sql.toString());
@@ -277,17 +240,12 @@ public class PermissionManager {
 			s.close();
 
 		} catch (Exception e) {
-			this.log
-					.error(
-							"Unexpected database error incurred in determining whether or not the user "
-									+ gridIdentity
-									+ "  is a trust service administrator, the following statement generated the error: \n"
-									+ sql.toString() + "\n", e);
+			this.log.error("Unexpected database error incurred in determining whether or not the user " + gridIdentity
+				+ "  is a trust service administrator, the following statement generated the error: \n"
+				+ sql.toString() + "\n", e);
 			GTSInternalFault fault = new GTSInternalFault();
-			fault
-					.setFaultString("Unexpected error occurred in determining whether or not the user "
-							+ gridIdentity
-							+ "  is a trust service administrator.");
+			fault.setFaultString("Unexpected error occurred in determining whether or not the user " + gridIdentity
+				+ "  is a trust service administrator.");
 			throw fault;
 		} finally {
 			db.releaseConnection(c);
@@ -295,8 +253,8 @@ public class PermissionManager {
 		return isAdmin;
 	}
 
-	public synchronized Permission[] findPermissions(PermissionFilter filter)
-			throws GTSInternalFault {
+
+	public synchronized Permission[] findPermissions(PermissionFilter filter) throws GTSInternalFault {
 
 		this.buildDatabase();
 		Connection c = null;
@@ -307,11 +265,9 @@ public class PermissionManager {
 			c = db.getConnection();
 			PreparedStatement s = null;
 			if (filter != null) {
-				s = c.prepareStatement("select * from "
-						+ PermissionsTable.TABLE_NAME + " WHERE "
-						+ PermissionsTable.GRID_IDENTITY + " LIKE ? AND "
-						+ PermissionsTable.ROLE + " LIKE ? AND "
-						+ PermissionsTable.TRUSTED_AUTHORITY + " LIKE ?");
+				s = c.prepareStatement("select * from " + PermissionsTable.TABLE_NAME + " WHERE "
+					+ PermissionsTable.GRID_IDENTITY + " LIKE ? AND " + PermissionsTable.ROLE + " LIKE ? AND "
+					+ PermissionsTable.TRUSTED_AUTHORITY + " LIKE ?");
 
 				if (filter.getGridIdentity() != null) {
 					s.setString(1, "%" + filter.getGridIdentity() + "%");
@@ -326,15 +282,12 @@ public class PermissionManager {
 				}
 
 				if (filter.getTrustedAuthorityName() != null) {
-					s
-							.setString(3, "%"
-									+ filter.getTrustedAuthorityName() + "%");
+					s.setString(3, "%" + filter.getTrustedAuthorityName() + "%");
 				} else {
 					s.setString(3, "%");
 				}
 			} else {
-				s = c.prepareStatement("select * from "
-						+ PermissionsTable.TABLE_NAME);
+				s = c.prepareStatement("select * from " + PermissionsTable.TABLE_NAME);
 			}
 
 			ResultSet rs = s.executeQuery();
@@ -342,8 +295,7 @@ public class PermissionManager {
 				Permission p = new Permission();
 				p.setGridIdentity(rs.getString(PermissionsTable.GRID_IDENTITY));
 				p.setRole(Role.fromValue(rs.getString(PermissionsTable.ROLE)));
-				p.setTrustedAuthorityName(clean(rs
-						.getString(PermissionsTable.TRUSTED_AUTHORITY)));
+				p.setTrustedAuthorityName(clean(rs.getString(PermissionsTable.TRUSTED_AUTHORITY)));
 				permissions.add(p);
 			}
 			rs.close();
@@ -356,18 +308,17 @@ public class PermissionManager {
 			return list;
 
 		} catch (Exception e) {
-			this.log
-					.error(
-							"Unexpected database error incurred in finding permissions, the following statement generated the error: \n"
-									+ sql.toString() + "\n", e);
+			this.log.error(
+				"Unexpected database error incurred in finding permissions, the following statement generated the error: \n"
+					+ sql.toString() + "\n", e);
 			GTSInternalFault fault = new GTSInternalFault();
-			fault
-					.setFaultString("Unexpected error occurred in finding permissions.");
+			fault.setFaultString("Unexpected error occurred in finding permissions.");
 			throw fault;
 		} finally {
 			db.releaseConnection(c);
 		}
 	}
+
 
 	private String clean(String s) {
 		if ((s == null) || (s.trim().length() == 0)) {
@@ -377,34 +328,34 @@ public class PermissionManager {
 		}
 	}
 
+
 	private String formatPermission(Permission p) {
 		String role = null;
 		if (p.getRole() != null) {
 			role = p.getRole().getValue();
 		}
-		return "[" + p.getGridIdentity() + "," + role + ","
-				+ p.getTrustedAuthorityName() + "]";
+		return "[" + p.getGridIdentity() + "," + role + "," + p.getTrustedAuthorityName() + "]";
 	}
+
 
 	public synchronized void buildDatabase() throws GTSInternalFault {
 		if (!dbBuilt) {
 			try {
 				db.createDatabase();
 				if (!this.db.tableExists(PermissionsTable.TABLE_NAME)) {
-					String sql = this.dbManager.getPermissionsTable()
-							.getCreateTableSQL();
+					String sql = this.dbManager.getPermissionsTable().getCreateTableSQL();
 					db.update(sql);
 				}
 				dbBuilt = true;
 			} catch (Exception e) {
 				this.log.error("Unexpected error in creating the database.", e);
 				GTSInternalFault fault = new GTSInternalFault();
-				fault
-						.setFaultString("Unexpected error in creating the database.");
+				fault.setFaultString("Unexpected error in creating the database.");
 				throw fault;
 			}
 		}
 	}
+
 
 	public synchronized void clearDatabase() throws GTSInternalFault {
 		try {
